feat(samples/action-center-app): act on review feedback, use published SDK 1.4.1
- Dedup shared task helpers (assignedToGroup, extractServerMessage) into taskUtils
- Detect group assignment via taskAssignmentCriteria (no cast); drop file:../.. for ^1.4.1
- Compute slaBadge once in TaskDetail; remove duplicate vite base key in process-app-v1
- Realign root oxlint lockfile to 1.43.0 to fix lint failures
