+++ samples/action-center-app/package.json
@@ -14,7 +14,7 @@
     "@uipath/apollo-wind": "^2.19.0",
     "@uipath/du-validation-station-wc": "1.0.0-beta.2",
     "@uipath/ui-widgets-validation-station": "^1.0.1-beta.6",
-    "@uipath/uipath-typescript": "file:../..",
+    "@uipath/uipath-typescript": "^1.4.1",
     "lucide-react": "^0.469.0",
     "next-themes": "^0.4.6",
     "path-browserify": "^1.0.1",
