+++ samples/process-app-v1/vite.config.ts
@@ -24,5 +24,4 @@ export default defineConfig({
   optimizeDeps: {
     include: ['@uipath/uipath-typescript'],
   },
-  base: './',
 })
